fix: localize perps twap labels and spot history direction
